Add pinned_comment field on publisher show method and comment text on…

+++ app/Http/Controllers/VideoController.php
@@ -481,24 +481,25 @@ class VideoController extends Controller
             }
 
 
+            if (!$video->published_at){
+                $pinnedComment = $video->pinned_comment;
 
-            $pinnedComment = $video->pinned_comment;
-
-            if($request->get('comment_text')){
-                if ($pinnedComment){
-                    $pinnedComment->text = $request->get('comment_text');
-                    $pinnedComment->save();
-                }else{
-                    $comment = new Comment();
-                    $comment->text = $request->get('comment_text')? : null;
-                    $comment->user_id = $video->user_id;
-                    $comment->is_pinned = Comment::COMMENT_PINNED;
-                    $comment->pinned_by = $video->user_id;
-                    $comment->video()->associate($video->id);
-                    $comment->save();
+                if($request->get('comment_text')){
+                    if ($pinnedComment){
+                        $pinnedComment->text = $request->get('comment_text');
+                        $pinnedComment->save();
+                    }else{
+                        $comment = new Comment();
+                        $comment->text = $request->get('comment_text')? : null;
+                        $comment->user_id = $video->user_id;
+                        $comment->is_pinned = Comment::COMMENT_PINNED;
+                        $comment->pinned_by = $video->user_id;
+                        $comment->video()->associate($video->id);
+                        $comment->save();
+                    }
+                }else if($pinnedComment){
+                    $pinnedComment->delete();
                 }
-            }else if($pinnedComment){
-                $pinnedComment->delete();
             }
 
         });
